Fix random property order in operator redo UI
Create global properties and assign them to operator props in every
operator, not ideal, but the only alternative I can think of is to use
one big operator instead of several small operators, which is more
complicated and harder to read. Or use class decorators, but that even
uglier than global props.